requirements.txt
# VRM Scraper - Runtime Dependencies
# Pinned versions for reproducibility

scrapy==2.11.0
openpyxl==3.1.2
PyYAML==6.0.1
